Move Rholang syntax file to root package
