+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/RholangCLI.scala
@@ -6,13 +6,11 @@ import cats.syntax.all._
 import coop.rchain.catscontrib.TaskContrib._
 import coop.rchain.metrics.{Metrics, NoopSpan, Span}
 import coop.rchain.models._
-import coop.rchain.rholang.interpreter.RhoRuntime.RhoISpace
 import coop.rchain.rholang.interpreter.accounting._
 import coop.rchain.rholang.interpreter.compiler.ParBuilder
 import coop.rchain.rholang.interpreter.errors._
 import coop.rchain.rholang.interpreter.storage.StoragePrinter
-import coop.rchain.rholang.interpreter.syntax._
-import coop.rchain.shared.Resources
+import coop.rchain.rholang.syntax._
 import coop.rchain.rspace.syntax.rspaceSyntaxKeyValueStoreManager
 import coop.rchain.shared.{Log, Resources}
 import coop.rchain.store.LmdbDirStoreManager.{mb, Db, LmdbEnvConfig}
