+++ rholang/src/test/scala/coop/rchain/rholang/InterpreterSpec.scala
@@ -7,8 +7,8 @@ import coop.rchain.models.{Expr, Par}
 import coop.rchain.rholang.Resources.mkRuntime
 import coop.rchain.rholang.interpreter.accounting._
 import coop.rchain.rholang.interpreter.storage.StoragePrinter
-import coop.rchain.rholang.interpreter.{EvaluateResult, Interpreter, RhoRuntime}
-import coop.rchain.rholang.interpreter.syntax._
+import coop.rchain.rholang.interpreter.{EvaluateResult, RhoRuntime}
+import coop.rchain.rholang.syntax._
 import coop.rchain.shared.Log
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
