+++ rholang/src/test/scala/coop/rchain/rholang/StackSafetySpec.scala
@@ -8,8 +8,8 @@ import coop.rchain.models._
 import coop.rchain.models.serialization.implicits._
 import coop.rchain.rholang.Resources.mkRuntime
 import coop.rchain.rholang.StackSafetySpec.findMaxRecursionDepth
-import coop.rchain.rholang.interpreter.{Interpreter, InterpreterUtil, ParBuilderUtil, PrettyPrinter}
-import coop.rchain.rholang.interpreter.syntax._
+import coop.rchain.rholang.interpreter.{ParBuilderUtil, PrettyPrinter}
+import coop.rchain.rholang.syntax._
 import coop.rchain.shared.{Log, Serialize}
 import monix.eval.{Coeval, Task}
 import monix.execution.Scheduler.Implicits.global
