+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/InterpreterUtil.scala
@@ -1,9 +1,8 @@
 package coop.rchain.rholang.interpreter
 
 import cats.effect.Sync
-import cats.implicits._
-import coop.rchain.models.Par
-import coop.rchain.rholang.interpreter.syntax._
+import cats.syntax.all._
+import coop.rchain.rholang.syntax._
 import org.scalatest.Matchers._
 
 object InterpreterUtil {
