+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/RuntimeSpec.scala
@@ -3,7 +3,7 @@ package coop.rchain.rholang.interpreter
 import coop.rchain.metrics
 import coop.rchain.metrics.{Metrics, NoopSpan, Span}
 import coop.rchain.rholang.Resources.mkRuntime
-import coop.rchain.rholang.interpreter.syntax._
+import coop.rchain.rholang.syntax._
 import coop.rchain.shared.Log
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
