+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/accounting/CostAccountingPropertyTest.scala
@@ -8,9 +8,9 @@ import coop.rchain.metrics.{Metrics, NoopSpan, Span}
 import coop.rchain.models._
 import coop.rchain.rholang.Resources._
 import coop.rchain.rholang.interpreter.{PrettyPrinter => PP, _}
-import coop.rchain.rholang.interpreter.syntax._
 import coop.rchain.rholang.ast.rholang_mercury.Absyn.{PPar, Proc}
 import coop.rchain.rholang.ast.rholang_mercury.PrettyPrinter
+import coop.rchain.rholang.syntax._
 import coop.rchain.rholang.{GenTools, ProcGen}
 import coop.rchain.shared.Log
 import monix.eval.{Coeval, Task}
