+++ rholang/src/test/scala/rholang/rosette/CompilerTests.scala
@@ -1,17 +1,16 @@
 package rholang.rosette
 
-import java.nio.file.{Files, Path, Paths}
-
 import coop.rchain.metrics
 import coop.rchain.metrics.{Metrics, NoopSpan, Span}
 import coop.rchain.rholang.Resources.mkRuntime
-import coop.rchain.rholang.interpreter.{EvaluateResult, Interpreter, InterpreterUtil}
-import coop.rchain.rholang.interpreter.syntax._
+import coop.rchain.rholang.interpreter.EvaluateResult
+import coop.rchain.rholang.syntax._
 import coop.rchain.shared.{Log, Resources}
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import org.scalatest.{FunSuite, Matchers}
 
+import java.nio.file.{Files, Path, Paths}
 import scala.collection.JavaConverters._
 import scala.concurrent.duration._
 import scala.io.Source
